Move elster request id check functions
